UXD-302 Fix: moved starling files
